abstracted out search
